Fixat feed controller lite

+++ BL/Controllers/CategoryController.cs
@@ -6,13 +6,18 @@ using DAL.Repositories;
 
 namespace BL.Controllers
 {
+   
+
     public class CategoryController
     {
+        Validering validator;
         ICategoryRepository<Category> categoryRepository;
 
         public CategoryController()
         {
             categoryRepository = new CategoryRepository();
+            validator = new Validering();
+
         }
         public void createCategory(string name)
         {
@@ -25,9 +30,18 @@ namespace BL.Controllers
         }
         public void deleteCategory(string category)
         {
-            int i = categoryRepository.GetIndexOfName(category);
-            categoryRepository.Delete(i);
+            if (validator.TextEmpty(category))
+            {
+                int i = categoryRepository.GetIndexOfName(category);
+                categoryRepository.Delete(i);
+            }
+            else
+            {
+                Console.WriteLine("No category");
+
+            }
         }
+            
 
         public void updatCategory(Category oldCat, string newCategoryName)
